[FEATURE] Load CSS files now inline in HTML-tags
This feature requires package tijsverkoyen/css-to-inline-styles to add styles inline to tags now. This is needed because webmailer often remove style-tags.
Sorting note: We support more then only one file. The first style for a tag is used in the package. Because this is completely different to normal TYPO3 template sorting, we just reverse the files before parsing through css-to-inline-styles package.

Related: https://github.com/in2code-de/luxletter/issues/102