[FEATURE] Multilanguage: Support localized container template names
This feature allows to automatically search for a specific language code in layout filename. E.g. if "NewsletterContainer.html" is defined in TypoScript, and the language 1 (de) should be used, luxletter automatically checks if there is a filename "NewsletterContainer_de.html".

In addition signal in constructor of NewsletterUrl was changed. Now only the class itself is passed. Getters and setters are available for origin, url and language. I did this because we have to replace signals with PSR-14 events in the near future and this is the first step.