[FEATURE] Show a better exception message if URL to unsubscribe page cannot be build
It turned out that we searched nearly 1h for a problem that mails could not be previewed on a client system. Source was that the given unsubscribe URL could not be build because of a non-existing PID. Now we want to show a useful message if this will happen again.

Related: https://projekte.in2code.de/issues/49756