[TASK] Use container layout names now without extension name .html
We used to pass layout names in backend module like "NewsletterContainer.html" in the past. But on a client system there was a htaccess redirect rule that redirects all links with something.html to something/. Of course this is not related to luxletter, but we are not sure if other users have the same problem out there.
We simply decided to remove the .html in the configuration in the same version when multilanguage support is released.

Related: https://projekte.in2code.de/issues/49755