[FEATURE] Allow to edit newsletter records
If an editor edits an existing newsletter record, all parameters can be changed but an existing queue is not touched. That means that you can change every meta information and even the bodytext is parsed again (to change newletter contents afterwards), but the receiver (no matter if already sent or not) list is not changed.