[BUGFIX] Fix user preview with sql_mode=only_full_group by
Errormessage: Expression #1 of SELECT list is not in GROUP BY clause and contains nonaggregated column 'typo3.fe_users.uid' which is not functionally dependent on columns in GROUP BY clause; this is incompatible with sql_mode=only_full_group_by

Related to default setting in MySQL only_full_group_by.
This was not that easy to fix, because we want to send only one newsletter per receiver. An email address can be used in more than only one fe_users record. In addition the fe_users record can also be related to two or even more usergroups that should receive a single newsletter.
So we need a grouping function to get single email addresses. And in a perfect world this would be done via SQL and not via PHP to prevent future performance and memory issues.
So I played and read a lot but I don't see a solution to still use "group by" even with any_value() or max() or such things. It's not possible to use "select * ... group by field" what is needed to build correct filled extbase models (with group by when only_full_group_by is set).
We now go for a grouping function in PHP - yes that's a pitty - to solve the issue. Because we changed QueryResult to array, more locations were updated with this change now

Fixes: https://github.com/in2code-de/luxletter/pull/143