[FEATURE] Also add unsubscribe URL in header as List-Unsubscribe value
to allow users to easily unsubscribe if header is supported in mail client.
This comment extracts unsubscribe url to a service class to be reused on two different cases now. We also return only a "#" for testmails and preview (because there is no newsletter-record yet).
We also introduced a new event UnsubscribeUrlEvent to be able to modify or overwrite the unsubscribe URL from outside.

Related: https://projekte.in2code.de/issues/55857