[FEATURE] First step to introduce a asynchronous queue feature
In this commit, we introduced a new switch to toggle the new feature in extension manager settings. In addition, new newsletters can be created without queue records and with a status tx_luxletter_domain_model_newsletter.queued=0. We also show a note in list view that those newsletters are not yet queued.
In the next step we need to add a commend to fill the queue afterwards.