[BUGFIX] Respect `asynchronousQueueStorage` if creating newsletter via CLI
The `asynchronousQueueStorage` option was introduced for large receiver volumes. The creation of the queue can take a long time here. Although CLI theoretically allows infinite script runtime, in the worst case a service/server restart can get in the way during this time and end the script run unexpectedly. For this reason, queues should also be set up asynchronously for newsletters created via CLI, if this has been configured accordingly in the instance.